chore: Move EventHub from :app to :core:eventhub
Precursor to other work that will use the eventhub from modules other
than :app.
