+++ app/build.gradle.kts
@@ -127,6 +127,7 @@ dependencies {
     implementation(projects.core.database)
     implementation(projects.core.designsystem)
     implementation(projects.core.domain)
+    implementation(projects.core.eventhub)
     implementation(projects.core.model)
     implementation(projects.core.navigation)
     implementation(projects.core.network)
