+++ app/src/main/java/app/pachli/appstore/CacheUpdater.kt
@@ -2,6 +2,14 @@ package app.pachli.appstore
 
 import app.pachli.core.data.repository.AccountManager
 import app.pachli.core.database.dao.TimelineDao
+import app.pachli.core.eventhub.BookmarkEvent
+import app.pachli.core.eventhub.EventHub
+import app.pachli.core.eventhub.FavoriteEvent
+import app.pachli.core.eventhub.PinEvent
+import app.pachli.core.eventhub.PollVoteEvent
+import app.pachli.core.eventhub.ReblogEvent
+import app.pachli.core.eventhub.StatusDeletedEvent
+import app.pachli.core.eventhub.UnfollowEvent
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
