+++ app/src/main/java/app/pachli/components/account/AccountViewModel.kt
@@ -3,12 +3,12 @@ package app.pachli.components.account
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.pachli.appstore.BlockEvent
-import app.pachli.appstore.DomainMuteEvent
-import app.pachli.appstore.EventHub
-import app.pachli.appstore.MuteEvent
-import app.pachli.appstore.ProfileEditedEvent
-import app.pachli.appstore.UnfollowEvent
+import app.pachli.core.eventhub.BlockEvent
+import app.pachli.core.eventhub.DomainMuteEvent
+import app.pachli.core.eventhub.EventHub
+import app.pachli.core.eventhub.MuteEvent
+import app.pachli.core.eventhub.ProfileEditedEvent
+import app.pachli.core.eventhub.UnfollowEvent
 import app.pachli.core.data.repository.AccountManager
 import app.pachli.core.network.model.Account
 import app.pachli.core.network.model.Relationship
