+++ app/src/main/java/app/pachli/components/conversation/ConversationsFragment.kt
@@ -37,7 +37,7 @@ import androidx.recyclerview.widget.SimpleItemAnimator
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout.OnRefreshListener
 import app.pachli.R
 import app.pachli.adapter.StatusBaseViewHolder
-import app.pachli.appstore.EventHub
+import app.pachli.core.eventhub.EventHub
 import app.pachli.core.activity.ReselectableFragment
 import app.pachli.core.common.extensions.hide
 import app.pachli.core.common.extensions.show
