+++ app/src/main/java/app/pachli/components/scheduled/ScheduledStatusActivity.kt
@@ -27,8 +27,8 @@ import androidx.lifecycle.lifecycleScope
 import androidx.paging.LoadState
 import androidx.recyclerview.widget.LinearLayoutManager
 import app.pachli.R
-import app.pachli.appstore.EventHub
-import app.pachli.appstore.StatusScheduledEvent
+import app.pachli.core.eventhub.EventHub
+import app.pachli.core.eventhub.StatusScheduledEvent
 import app.pachli.core.activity.BaseActivity
 import app.pachli.core.common.extensions.hide
 import app.pachli.core.common.extensions.show
